make info api logic reusable
